requirements.txt
pyccl~=2.0
numpy>=1.16
pyyaml
scipy
click
pandas
jinja2
schwimmbad
emcee>=3,<5
tqdm
joblib
sacc~=0.2
